organize i18n config files
